backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.19
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
cryptography==42.0.5
proxmoxer==2.0.1
requests==2.31.0
pydantic-settings==2.2.1
prometheus-fastapi-instrumentator==6.1.0
